:truck: move: move home component to container directory
