+++ knutice/src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as FirstSection } from './Main/FirstSection';
 export { default as SecondSection } from './Main/SecondSection';
-export { default as Home } from './Main/Home';
+export { default as Home } from './Container/Home';
 
 export { default as Navbar } from './Navbar/Navbar';
